Update pi benchmark test and add feature
